Add view pager to tabs in notifications list fragment
