+++ WordPress/src/main/res/layout/notifications_list_fragment.xml
@@ -7,62 +7,13 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent">
 
-    <org.wordpress.android.util.widgets.CustomSwipeRefreshLayout
-        android:id="@+id/swipe_refresh_notifications"
+    <org.wordpress.android.widgets.WPViewPager
+        android:id="@+id/view_pager"
         android:layout_below="@+id/app_bar"
         android:layout_height="match_parent"
         android:layout_width="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/recycler_view_notes"
-            android:clipToPadding="false"
-            android:layout_height="match_parent"
-            android:layout_width="match_parent"
-            android:paddingEnd="@dimen/notifications_content_margin"
-            android:paddingStart="@dimen/notifications_content_margin"
-            android:scrollbars="vertical"
-            android:scrollbarStyle="outsideOverlay"
-            tools:listitem="@layout/notifications_list_item">
-        </android.support.v7.widget.RecyclerView>
-
-    </org.wordpress.android.util.widgets.CustomSwipeRefreshLayout>
-
-    <LinearLayout
-        android:id="@+id/layout_new_notificatons"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="bottom|center_horizontal"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="24dp"
-        android:layout_width="wrap_content"
-        android:background="@drawable/notifications_new_notification_bar_background"
-        android:elevation="@dimen/message_bar_elevation"
-        android:orientation="vertical"
-        android:paddingBottom="@dimen/margin_large"
-        android:paddingStart="24dp"
-        android:paddingEnd="24dp"
-        android:paddingTop="@dimen/margin_large"
-        android:visibility="gone"
-        tools:visibility="visible">
-
-        <org.wordpress.android.widgets.WPTextView
-            android:layout_height="wrap_content"
-            android:layout_width="wrap_content"
-            android:text="@string/notifications_label_new_notifications"
-            android:textColor="@color/grey_lighten_30"
-            android:textSize="@dimen/text_sz_large">
-        </org.wordpress.android.widgets.WPTextView>
-
-        <org.wordpress.android.widgets.WPTextView
-            android:layout_height="wrap_content"
-            android:layout_width="wrap_content"
-            android:text="@string/notifications_label_new_notifications_subtitle"
-            android:textColor="@color/grey_lighten_30"
-            android:textSize="@dimen/text_sz_small">
-        </org.wordpress.android.widgets.WPTextView>
-
-    </LinearLayout>
+    </org.wordpress.android.widgets.WPViewPager>
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
@@ -98,19 +49,6 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <org.wordpress.android.ui.ActionableEmptyView
-        android:id="@+id/actionable_empty_view"
-        android:layout_height="match_parent"
-        android:layout_marginTop="@dimen/toolbar_height"
-        android:layout_width="match_parent"
-        android:visibility="gone"
-        app:aevButton="@string/posts_empty_list_button"
-        app:aevImage="@drawable/img_illustration_notifications_152dp"
-        app:aevSubtitle="@string/notifications_empty_action_unread"
-        app:aevTitle="@string/notifications_empty_unread"
-        tools:visibility="visible">
-    </org.wordpress.android.ui.ActionableEmptyView>
-
     <LinearLayout
         android:id="@+id/connect_jetpack"
         android:gravity="center"
@@ -118,7 +56,8 @@
         android:layout_marginTop="@dimen/toolbar_height"
         android:layout_width="match_parent"
         android:orientation="vertical"
-        android:visibility="invisible">
+        android:visibility="invisible"
+        tools:visibility="visible">
 
         <android.support.v4.widget.NestedScrollView
             android:layout_height="wrap_content"
